refactor: entities services to entity directory
Took 2 minutes
