+++ src/main/java/dev/realtards/kuenyawz/controllers/ProductController.java
@@ -2,8 +2,8 @@ package dev.realtards.kuenyawz.controllers;
 
 import dev.realtards.kuenyawz.dtos.product.*;
 import dev.realtards.kuenyawz.services.ProductCsvService;
-import dev.realtards.kuenyawz.services.ProductService;
-import dev.realtards.kuenyawz.services.VariantService;
+import dev.realtards.kuenyawz.services.entity.ProductService;
+import dev.realtards.kuenyawz.services.entity.VariantService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
