+++ src/main/java/dev/realtards/kuenyawz/services/ProductCsvServiceImpl.java
@@ -6,6 +6,7 @@ import dev.realtards.kuenyawz.dtos.product.ProductPostDto;
 import dev.realtards.kuenyawz.dtos.product.VariantPostDto;
 import dev.realtards.kuenyawz.exceptions.InvalidRequestBodyValue;
 import dev.realtards.kuenyawz.exceptions.ResourceUploadException;
+import dev.realtards.kuenyawz.services.entity.ProductService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
